requirements.txt
torch
numpy
tqdm
pillow
transformers
ipykernel
safetensors